Flyttet font-håndtering til pdfutils
